Get only the initially-visible post at page load
Previous versions created a separate iframe for each post and hid all
but the currently-selected one with CSS. This caused browsers to fetch
the contents for all posts. I wrote a little JavaScript that tried to
stop the iframes from loading but it raced with the browser so some
subset of the iframes would have already loaded by the time the script
ran.

This version instead requires JavaScript to make the top bar and side
bar visually reflect which post is currently loaded, although selecting
posts in the left side bar works even without JavaScript.