Use a pure-HTML interface
This removes the XSLT templating in favor of server-side templating as
is standard in other web development, and removes all need for
JavaScript.

The downside is that this version requires sending a request to the
server for every navigation.

I could add some JavaScript that, if enabled, avoids hitting the server,
while falling back to the pure-HTML approach when disabled. That might
be fun later.

For now, I've just tried to ensure that the server can send as small of
a response as possible for each navigation, and that downstream caches
are as effective as possible.